greeter_web/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.2.1
grpcio==1.48.0
grpcio-tools==1.48.0
idna==3.3
itsdangerous==2.1.2
jaeger-client==4.8.0
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
nacos-sdk-python==0.1.8
opencv-contrib-python==4.6.0.66
opentracing==2.4.0
protobuf==3.20.1
pybreaker==1.0.1
python-consul2==0.1.5
ratelimit==2.2.1
requests==2.28.1
six==1.16.0
threadloop==1.0.2
thrift==0.16.0
tornado==6.2
urllib3==1.26.11
Werkzeug==2.2.1
win32-setctime==1.1.0
